Add files via upload
-'alpha3D.m' program solves the 3D Schrôdinger's equation for the coulombian potential, since it's analitically solvable. It works for other potentials by using the finite difference method as an approximation.
-'alphaWKB.m' uses the WKB approximation to solve the equation for any potential. It is used to compare the results with the previous method.
-'qpAlphaDecayEAN.m' is used to plot the kinetic energy vs the atomic number N for different isotopes.

The whole plot is about solving the 3D schrôdinger's equation to calculate the half lifes of the many isotopes we can introduce as inputs. Then, we compare the results with the real or measured ones. We also calculate the errors to get an idea about which method is best to solve the equation to obtain the  closer results compared to the measured ones.